==> notebooks/preprocessing/01_compute_rolling_background_for_fluorescence_data.ipynb <==
%load_ext autoreload
%autoreload 2
%matplotlib inline

import os
import numpy as np
import glob
import h5py
from tqdm.auto import tqdm
from markovids import vid
from joblib import Parallel, delayed

## User functions

def get_rolling_bground(
    dat_file,
    step_size=1500,
    frame_spacing=20,
    overlap=0,
    agg_func=np.mean,
    reader_kwargs={"threads": 2},
    progress_bar=False,
    save_dir="_bground",
):

    basename = os.path.splitext(os.path.basename(dat_file))[0]
    path = os.path.dirname(dat_file)
    bground_path = os.path.join(path, save_dir, f"{basename}.hdf5")

    os.makedirs(os.path.join(path, save_dir), exist_ok=True)

    if os.path.exists(bground_path):
        return None

    vid_reader = vid.io.AutoReader(dat_file, **reader_kwargs)
    steps = range(0, vid_reader.nframes, step_size - overlap)

    bgrounds = []
    idxs = []
    for _step in tqdm(steps, disable=not progress_bar):
        left_edge = _step
        right_edge = left_edge + step_size

        if right_edge > vid_reader.nframes:
            break
        # right_edge = min(left_edge + step_size, reader.nframes)
        use_frames = vid_reader.get_frames(range(left_edge, right_edge, frame_spacing))
        bgrounds.append(agg_func(use_frames, axis=0))
        idxs.append((left_edge + right_edge) / 2)  # get midpoint frame index...
    vid_reader.close()

    with h5py.File(bground_path, "w") as f:
        f.create_dataset(
            "bground",
            data=np.array(bgrounds),
            compression="gzip",
        )
        f.create_dataset("frame_idxs", data=np.array(idxs), compression="gzip")

    return None

# Quantify fluorescence decay over time

base_dir = os.path.expanduser("~/shared_folder/active_lab_members/markowitz_jeffrey/active_projects/quantum_dots/timecourse_0*")
fluo_files = sorted(glob.glob(os.path.join(base_dir, "**", "Basler*fluorescence.avi"), recursive=True))

base_dir = os.path.expanduser("~/shared_folder/active_lab_members/markowitz_jeffrey/active_projects/quantum_dots/sciadv_rebuttal/")
fluo_files += sorted(glob.glob(os.path.join(base_dir, "**", "Basler*fluorescence.avi"), recursive=True))

delays = []
for _file in fluo_files:
    delays.append(
        delayed(get_rolling_bground)(
            _file,
        )
    )
print(len(delays))
dat = Parallel(n_jobs=8, verbose=10, batch_size=8, backend="multiprocessing")(delays)
