==> notebooks/preprocessing/10_convert_slp_to_embedded.ipynb <==
import glob
import os
from tqdm.auto import tqdm


import shutil

# file_filter = "/storage/home/hcoda1/4/jmarkowitz30/shared_folder/active_lab_members/markowitz_jeffrey/active_projects/keypoints_basler_nir_plexiglass_arena/sleap_training_round2/_labels_qd/*embed-False*.slp" 

file_filter = "/storage/home/hcoda1/4/jmarkowitz30/shared_folder/active_lab_members/markowitz_jeffrey/active_projects/keypoints_basler_nir_plexiglass_arena/sleap_training_round2/_labels_qd/kpoint_training_dataset_qd_alignment_kneejoints_reflect_only_round2_manual_labeling_nofluo_version-v[1-2]-embed-False.slp"

data_files = sorted(glob.glob(file_filter))

print(data_files)

force = False

# sleap_io version

import sleap_io as sio

for original_file in tqdm(data_files):
    original_mdata_file = original_file.replace(".slp", ".toml")
    new_file = original_file.replace("embed-False", "embed-True")
    if os.path.exists(new_file) and (not force):
        continue
    new_mdata_file = new_file.replace(".slp", ".toml")
    original_data = sio.load_file(original_file)
    original_data.save(new_file, embed=True)
    shutil.copyfile(original_mdata_file, new_mdata_file)

# SLEAP version (for some reason only this works for large embedded files)

import sleap

for original_file in tqdm(data_files):
    original_mdata_file = original_file.replace(".slp", ".toml")
    new_file = original_file.replace("embed-False", "embed-True-sleap-version")
    if os.path.exists(new_file) and (not force):
        continue
    new_mdata_file = new_file.replace(".slp", ".toml")
    original_data = sleap.load_file(original_file)
    original_data.save(new_file, with_images=True)
    shutil.copyfile(original_mdata_file, new_mdata_file)

print(new_file)
